+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - VirtualDAC now has an option to set gate_map and instruments on VirtualDAC (#583).
 - All unittests are moved to a tests folder (#574).
 - Legacy code now throws exceptions (#551).
+- VideoMode tool was restructured to clean up the interface (##594)
 
 ### Deprecated
 - loadExperimentData and saveExperimentData are now deprecated functions (#581).
@@ -36,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ...
 
 ### Fixed
+- Fixed virtual AWG marker on the same channel (#599).
 - Fixed the readthedocs build (#590).
 - Fixed issue with HDAWG8 sampling rate (#589).
 - Fixed Travis warnings and tests instrument_drivers (#569).